Moved auditing and integration utils from the web content service to the common.
