+++ abixen-platform-common/src/main/java/com/abixen/platform/common/client/SecurityClient.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.abixen.platform.service.webcontent.client;
+package com.abixen.platform.common.client;
 
-import com.abixen.platform.service.webcontent.model.User;
+import com.abixen.platform.common.model.enumtype.AclClassName;
 import org.springframework.cloud.netflix.feign.FeignClient;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,9 +22,14 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 
 @FeignClient("abixen-platform-core")
-public interface UserClient {
+public interface SecurityClient {
 
-    @RequestMapping(method = RequestMethod.GET, value = "/api-intranet/users/{id}")
-    User getUserById(@PathVariable("id") Long id);
+    @RequestMapping(method = RequestMethod.GET,
+            value = "/api/control-panel/securities/has-permission/{username}/{securableObjectId}/{aclClassName}/{permissionName}")
+    Boolean hasPermission(@PathVariable("username") String username,
+                          @PathVariable("securableObjectId") Long securableObjectId,
+                          @PathVariable("aclClassName") AclClassName aclClassName,
+                          @PathVariable("permissionName") String permissionName);
 
 }
+
