+++ abixen-platform-web-content-service/src/main/java/com/abixen/platform/service/webcontent/converter/TemplateToTemplateDtoConverter.java
@@ -14,6 +14,7 @@
 package com.abixen.platform.service.webcontent.converter;
 
 import com.abixen.platform.common.converter.AbstractConverter;
+import com.abixen.platform.common.converter.AuditingModelToAuditingDtoConverter;
 import com.abixen.platform.service.webcontent.dto.TemplateDto;
 import com.abixen.platform.service.webcontent.model.impl.Template;
 import org.springframework.beans.factory.annotation.Autowired;
