+++ abixen-platform-web-content-service/src/main/java/com/abixen/platform/service/webcontent/converter/WebContentModuleConfigurationToWebContentModuleConfigurationDtoConverter.java
@@ -14,6 +14,7 @@
 package com.abixen.platform.service.webcontent.converter;
 
 import com.abixen.platform.common.converter.AbstractConverter;
+import com.abixen.platform.common.converter.AuditingModelToAuditingDtoConverter;
 import com.abixen.platform.service.webcontent.dto.WebContentModuleConfigurationDto;
 import com.abixen.platform.service.webcontent.model.impl.WebContentModuleConfig;
 import org.springframework.beans.factory.annotation.Autowired;
