+++ abixen-platform-web-content-service/src/main/java/com/abixen/platform/service/webcontent/converter/WebContentToWebContentDtoConverter.java
@@ -14,6 +14,7 @@
 package com.abixen.platform.service.webcontent.converter;
 
 import com.abixen.platform.common.converter.AbstractConverter;
+import com.abixen.platform.common.converter.AuditingModelToAuditingDtoConverter;
 import com.abixen.platform.service.webcontent.dto.AdvancedWebContentDto;
 import com.abixen.platform.service.webcontent.dto.SimpleWebContentDto;
 import com.abixen.platform.service.webcontent.dto.StructureDto;
