+++ abixen-platform-web-content-service/src/main/java/com/abixen/platform/service/webcontent/security/PlatformPermissionEvaluator.java
@@ -14,9 +14,9 @@
 
 package com.abixen.platform.service.webcontent.security;
 
+import com.abixen.platform.common.integration.SecurityIntegrationClient;
 import com.abixen.platform.common.model.enumtype.AclClassName;
 import com.abixen.platform.common.security.PlatformUser;
-import com.abixen.platform.service.webcontent.integration.SecurityIntegrationClient;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.NotImplementedException;
 import org.springframework.beans.factory.annotation.Autowired;
